move to  NextLedgerUpdateCommitment
